DEV: Consolidate chatbot implementation under `Bot`
